src/integ_test_resources/ios/sdk/integration/cdk/requirements.txt
aws-cdk-lib >= 2.0.0
constructs >= 10.0.0
boto3
